+++ app/src/main/java/projects/medicationtracker/Dialogs/AddAsNeededDoseDialog.java
@@ -26,7 +26,7 @@ import java.util.stream.Collectors;
 
 import projects.medicationtracker.Fragments.TimePickerFragment;
 import projects.medicationtracker.Helpers.DBHelper;
-import projects.medicationtracker.Helpers.TimeFormatting;
+import projects.medicationtracker.Utils.TimeFormatting;
 import projects.medicationtracker.Interfaces.IDialogCloseListener;
 import projects.medicationtracker.R;
 import projects.medicationtracker.Models.Dose;
