+++ app/src/main/java/projects/medicationtracker/Fragments/MedicationScheduleFragment.java
@@ -40,8 +40,8 @@ import projects.medicationtracker.Dialogs.AddAsNeededDoseDialog;
 import projects.medicationtracker.Dialogs.DoseInfoDialog;
 import projects.medicationtracker.Helpers.DBHelper;
 import projects.medicationtracker.Helpers.NativeDbHelper;
-import projects.medicationtracker.Helpers.TextViewUtils;
-import projects.medicationtracker.Helpers.TimeFormatting;
+import projects.medicationtracker.Utils.TextViewUtils;
+import projects.medicationtracker.Utils.TimeFormatting;
 import projects.medicationtracker.Interfaces.IDialogCloseListener;
 import projects.medicationtracker.R;
 import projects.medicationtracker.Models.Dose;
