+++ app/src/main/java/projects/medicationtracker/Models/Medication.java
@@ -12,7 +12,7 @@ import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Locale;
 
-import projects.medicationtracker.Helpers.TimeFormatting;
+import projects.medicationtracker.Utils.TimeFormatting;
 import projects.medicationtracker.R;
 
 public class Medication implements Cloneable, Parcelable {
