+++ app/src/main/java/projects/medicationtracker/Receivers/ExportReceiver.java
@@ -3,11 +3,12 @@ package projects.medicationtracker.Receivers;
 import android.content.BroadcastReceiver;
 import android.content.Context;
 import android.content.Intent;
+import android.os.Bundle;
 
 public class ExportReceiver extends BroadcastReceiver {
 
     @Override
     public void onReceive(Context context, Intent intent) {
-
+        Bundle extras = intent.getExtras();
     }
 }
