+++ app/src/main/java/projects/medicationtracker/Utils/DataExportUtils.java
@@ -1,0 +1,9 @@
+package projects.medicationtracker.Utils;
+
+import java.time.LocalDateTime;
+
+public class DataExportUtils {
+    public static void scheduleExport(LocalDateTime exportTime, int frequency) {
+
+    }
+}
